requirements.txt
omegaconf~=2.3.0
kedro~=0.19.6
uvicorn~=0.30.1
fastapi~=0.111.0
pydantic~=2.7.3
pytest~=8.2.2
requests~=2.32.3
pandas~=2.0.3
click~=8.1.7
cookiecutter~=2.6.0
PyYAML>=4.2,<7.0